+++ src/hooks/useProducts.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import type { Product } from '../product.type';
+import type { Product } from '../types/product.type';
 import { apiClient } from '../api-client';
 
 export function useProducts() {
@@ -11,6 +11,7 @@ export function useProducts() {
             setIsLoading(true);
             try {
                 const response = await apiClient.get('/products');
+                console.log('response: ', response);
                 setProducts(response.data.data);
             } catch (error) {
                 console.error(error);
@@ -20,7 +21,6 @@ export function useProducts() {
         };
 
         fetchProducts();
-        return () => {};
     }, []);
 
     return { products, isLoading };
